requirements.txt
Django==2.2.6
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
mysqlclient==1.4.4
Pillow==6.2.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
